+++ tests/nn/data_parallel/test_sharded_ddp_features.py
@@ -22,7 +22,6 @@ from fairscale.optim import OSS
 from fairscale.utils.testing import (
     GPT2,
     SGDWithPausingCompute,
-    available_devices,
     check_same_models_across_ranks,
     skip_if_less_than_four_gpu,
     skip_if_no_cuda,
@@ -46,19 +45,13 @@ class _DoubleInput(torch.nn.Module):
 
 
 def run_one_step(
-    rank,
-    world_size,
-    backend,
-    device,
-    temp_file_name,
-    broadcast_buffers,
-    grad_accumulation,
-    reduce_buffer_size,
-    optimizer_type,
+    rank, world_size, temp_file_name, broadcast_buffers, grad_accumulation, reduce_buffer_size, optimizer_type,
 ):
+    device = torch.device("cuda")
+    backend = dist.Backend.NCCL
+
     dist.init_process_group(init_method="file://" + temp_file_name, backend=backend, rank=rank, world_size=world_size)
-    if device == torch.device("cuda"):
-        torch.cuda.set_device(rank)
+    torch.cuda.set_device(rank)
 
     torch.manual_seed(rank)
     np.random.seed(rank)
@@ -115,36 +108,10 @@ def run_test(backend, device, world_size, broadcast_buffers, grad_accumulation, 
     temp_file_name = tempfile.mkstemp()[1]
     mp.spawn(
         run_one_step,
-        args=(world_size, backend, device, temp_file_name, broadcast_buffers, grad_accumulation, reduce_buffer_size),
-        nprocs=world_size,
-        join=True,
-    )
-
-
-@skip_if_no_cuda
-@skip_if_single_gpu
-@pytest.mark.parametrize("broadcast_buffers", [True, False])
-@pytest.mark.parametrize("grad_accumulation", [True, False])
-@pytest.mark.parametrize("reduce_buffer_size", [0, 2 ** 20])
-@pytest.mark.parametrize("optimizer_type", [torch.optim.SGD, SGDWithPausingCompute])
-@pytest.mark.parametrize(
-    "setup",
-    [
-        [dist.Backend.NCCL, torch.device("cuda")],
-        [dist.Backend.GLOO, torch.device("cpu")],
-        [dist.Backend.GLOO, torch.device("cuda")],
-    ],
-)
-def test_step(broadcast_buffers, grad_accumulation, reduce_buffer_size, optimizer_type, setup):
-    world_size = 2
-    temp_file_name = tempfile.mkstemp()[1]
-
-    mp.spawn(
-        run_one_step,
         args=(
             world_size,
-            setup[0],
-            setup[1],
+            backend,
+            device,
             temp_file_name,
             broadcast_buffers,
             grad_accumulation,
@@ -156,10 +123,29 @@ def test_step(broadcast_buffers, grad_accumulation, reduce_buffer_size, optimize
     )
 
 
-def run_test_two_inputs(rank, world_size, backend, device, temp_file_name, reduce_buffer_size):
+@skip_if_no_cuda
+@skip_if_single_gpu
+@pytest.mark.parametrize("broadcast_buffers", [True, False])
+@pytest.mark.parametrize("grad_accumulation", [True, False])
+@pytest.mark.parametrize("reduce_buffer_size", [0, 2 ** 20])
+@pytest.mark.parametrize("optimizer_type", [torch.optim.SGD, SGDWithPausingCompute])
+def test_step(broadcast_buffers, grad_accumulation, reduce_buffer_size, optimizer_type):
+    world_size = 2
+    temp_file_name = tempfile.mkstemp()[1]
+
+    mp.spawn(
+        run_one_step,
+        args=(world_size, temp_file_name, broadcast_buffers, grad_accumulation, reduce_buffer_size, optimizer_type,),
+        nprocs=world_size,
+        join=True,
+    )
+
+
+def run_test_two_inputs(rank, world_size, temp_file_name, reduce_buffer_size):
+    device = torch.device("cuda")
+    backend = dist.Backend.NCCL
     dist.init_process_group(init_method="file://" + temp_file_name, backend=backend, rank=rank, world_size=world_size)
-    if device == "cuda":
-        torch.cuda.set_device(rank)
+    torch.cuda.set_device(rank)
 
     torch.manual_seed(rank)
     np.random.seed(rank)
@@ -183,47 +169,45 @@ def run_test_two_inputs(rank, world_size, backend, device, temp_file_name, reduc
 
 
 @pytest.mark.parametrize("reduce_buffer_size", [0, 2 ** 20])
-@pytest.mark.parametrize("backend", ["gloo", "nccl"])
-@pytest.mark.parametrize("device", available_devices)
-def test_inputs(reduce_buffer_size, backend, device):
+@skip_if_no_cuda
+def test_inputs(reduce_buffer_size):
     # Check that the ShardedDDP wrapper accepts tuple(tensors) as inputs
     world_size = 2
-    if backend == "nccl" and device == "cpu":
-        pytest.skip("Incompatible combination, or cuda not available")
-        return
 
     mp.spawn(
-        run_test_two_inputs,
-        args=(world_size, backend, device, tempfile.mkstemp()[1], reduce_buffer_size),
-        nprocs=world_size,
-        join=True,
+        run_test_two_inputs, args=(world_size, tempfile.mkstemp()[1], reduce_buffer_size), nprocs=world_size, join=True,
     )
 
 
+@skip_if_no_cuda
 def test_ddp_attributes():
     # Check that ShardedDDP exposes the same attributes as Pytorch's DDP
     # - is multi_device_module
     # - device_type
-    dist.init_process_group(init_method="file://" + tempfile.mkstemp()[1], backend="gloo", rank=0, world_size=1)
+    backend = dist.Backend.NCCL
+    dist.init_process_group(init_method="file://" + tempfile.mkstemp()[1], backend=backend, rank=0, world_size=1)
 
     model = Sequential(Linear(2, 3), Linear(3, 3))
     optimizer = OSS(params=model.parameters(), optim=torch.optim.SGD, lr=1e-3, momentum=0.99)
-    ddp_model = ShardedDataParallel(model, optimizer)
+    ddp_model = ShardedDataParallel(model, optimizer, sync_models_at_startup=False)
 
     assert hasattr(ddp_model, "is_multi_device_module")
     assert hasattr(ddp_model, "device_type")
     dist.destroy_process_group()
 
 
+@skip_if_no_cuda
 def test_random_attributes():
+    backend = dist.Backend.NCCL
+
     # Check that ShardedDDP exposes the original module's attributes
-    dist.init_process_group(init_method="file://" + tempfile.mkstemp()[1], backend="gloo", rank=0, world_size=1)
+    dist.init_process_group(init_method="file://" + tempfile.mkstemp()[1], backend=backend, rank=0, world_size=1)
 
     model = Sequential(Linear(2, 3), Linear(3, 3))
     model.banana = "sweet"
 
     optimizer = OSS(params=model.parameters(), optim=torch.optim.SGD, lr=1e-3, momentum=0.99)
-    ddp_model = ShardedDataParallel(model, optimizer)
+    ddp_model = ShardedDataParallel(model, optimizer, sync_models_at_startup=False)
 
     assert hasattr(ddp_model, "banana")
     assert not hasattr(ddp_model, "orange")
@@ -231,8 +215,10 @@ def test_random_attributes():
     dist.destroy_process_group()
 
 
-def run_test_device_change(rank, world_size, backend, device, temp_file_name, reduce_buffer_size):
+def run_test_device_change(rank, world_size, temp_file_name, reduce_buffer_size):
     # Check that the wrapped module can change devices
+    device = torch.device("cuda")
+    backend = dist.Backend.NCCL
     dist.init_process_group(init_method="file://" + temp_file_name, backend=backend, rank=rank, world_size=world_size)
     torch.cuda.set_device(rank)
 
@@ -256,18 +242,15 @@ def run_test_device_change(rank, world_size, backend, device, temp_file_name, re
 def test_device_change(reduce_buffer_size):
     # Check that ShardedDDP is compatible with sync batch norm across multiple GPUs
     world_size = 2
-    backend = "nccl"
     temp_file_name = tempfile.mkstemp()[1]
-    device = "cuda"
     mp.spawn(
-        run_test_device_change,
-        args=(world_size, backend, device, temp_file_name, reduce_buffer_size),
-        nprocs=world_size,
-        join=True,
+        run_test_device_change, args=(world_size, temp_file_name, reduce_buffer_size), nprocs=world_size, join=True,
     )
 
 
-def run_test_training_change(rank, world_size, backend, device, temp_file_name, reduce_buffer_size):
+def run_test_training_change(rank, world_size, temp_file_name, reduce_buffer_size):
+    device = torch.device("cuda")
+    backend = dist.Backend.NCCL
     group = dist.init_process_group(
         init_method="file://" + temp_file_name, backend=backend, rank=rank, world_size=world_size
     )
@@ -293,23 +276,20 @@ def run_test_training_change(rank, world_size, backend, device, temp_file_name, 
 @pytest.mark.parametrize("reduce_buffer_size", [0, 2 ** 20])
 def test_training_change(reduce_buffer_size):
     world_size = 2
-    backend = "nccl"
     temp_file_name = tempfile.mkstemp()[1]
-    device = "cuda"
     mp.spawn(
-        run_test_training_change,
-        args=(world_size, backend, device, temp_file_name, reduce_buffer_size),
-        nprocs=world_size,
-        join=True,
+        run_test_training_change, args=(world_size, temp_file_name, reduce_buffer_size), nprocs=world_size, join=True,
     )
 
 
-def run_test_ddp_sync_batch_norm(rank, world_size, backend, device, temp_file_name):
+def run_test_ddp_sync_batch_norm(rank, world_size, temp_file_name):
+    device = torch.device("cuda")
+    backend = dist.Backend.NCCL
+    torch.cuda.set_device(rank)
     dist.init_process_group(init_method="file://" + temp_file_name, backend=backend, rank=rank, world_size=world_size)
 
     model = Sequential(Linear(2, 3), torch.nn.BatchNorm1d(3), Linear(3, 3)).to(device)
     model = torch.nn.SyncBatchNorm.convert_sync_batchnorm(model)
-    model.to(device)  # in pytorch 1.5 syncBN switches to the default device/cpu
 
     optimizer = OSS(params=model.parameters(), optim=torch.optim.SGD, lr=1e-3, momentum=0.99)
     ddp_model = ShardedDataParallel(model, optimizer)
@@ -325,18 +305,15 @@ def run_test_ddp_sync_batch_norm(rank, world_size, backend, device, temp_file_na
 def test_ddp_sync_batch_norm():
     # Check that ShardedDDP is compatible with sync batch norm across multiple GPUs
     world_size = 2
-    backend = "gloo"
     temp_file_name = tempfile.mkstemp()[1]
-    device = "cuda"
-    mp.spawn(
-        run_test_ddp_sync_batch_norm, args=(world_size, backend, device, temp_file_name), nprocs=world_size, join=True
-    )
+    mp.spawn(run_test_ddp_sync_batch_norm, args=(world_size, temp_file_name), nprocs=world_size, join=True)
 
 
-def run_test_two_optimizers(rank, world_size, backend, device, temp_file_name):
+def run_test_two_optimizers(rank, world_size, temp_file_name):
+    device = torch.device("cuda")
+    backend = dist.Backend.NCCL
     dist.init_process_group(init_method="file://" + temp_file_name, backend=backend, rank=rank, world_size=world_size)
-    if device == torch.device("cuda"):
-        torch.cuda.set_device(rank)
+    torch.cuda.set_device(rank)
 
     torch.manual_seed(rank)
     np.random.seed(rank)
@@ -364,21 +341,23 @@ def run_test_two_optimizers(rank, world_size, backend, device, temp_file_name):
     dist.destroy_process_group()
 
 
+@skip_if_no_cuda
+@skip_if_single_gpu
 def test_two_optimizers():
     # Check that the ShardedDDP wrapper accepts tuple(tensors) as inputs
     world_size = 2
-    backend = "gloo"
     temp_file_name = tempfile.mkstemp()[1]
-    device = "cpu"
-    mp.spawn(run_test_two_optimizers, args=(world_size, backend, device, temp_file_name), nprocs=world_size, join=True)
+    mp.spawn(run_test_two_optimizers, args=(world_size, temp_file_name), nprocs=world_size, join=True)
 
 
-def run_test_gpt2(rank, world_size, backend, device, temp_file_name):
+def run_test_gpt2(rank, world_size, temp_file_name):
     INPUT_DIM = 16
     BACH_SIZE = 10
     STEPS = 10
 
     url = "file://" + temp_file_name
+    device = torch.device("cuda")
+    backend = dist.Backend.NCCL
     dist.init_process_group(init_method=url, backend=backend, rank=rank, world_size=world_size)
     torch.cuda.set_device(rank)
 
@@ -411,13 +390,14 @@ def run_test_gpt2(rank, world_size, backend, device, temp_file_name):
 @pytest.mark.parametrize("world_size", [1, 2])
 def test_gpt2(world_size):
     # Check that having trainable unused params is fine
-    backend = "gloo"
     temp_file_name = tempfile.mkstemp()[1]
-    device = "cuda"
-    mp.spawn(run_test_gpt2, args=(world_size, backend, device, temp_file_name), nprocs=world_size, join=True)
+    mp.spawn(run_test_gpt2, args=(world_size, temp_file_name), nprocs=world_size, join=True)
 
 
-def run_test_multiple_groups(rank, world_size, tempfile_name, backend, reduce_buffer_size):
+def run_test_multiple_groups(rank, world_size, tempfile_name, reduce_buffer_size):
+    device = torch.device("cuda")
+    backend = dist.Backend.NCCL
+
     # Only work with the even ranks, to check that the global_rank indexing is properly used
     dist.init_process_group(init_method="file://" + tempfile_name, backend=backend, rank=rank, world_size=world_size)
 
@@ -430,7 +410,6 @@ def run_test_multiple_groups(rank, world_size, tempfile_name, backend, reduce_bu
     np.random.seed(rank)
 
     # Standard deep learning setup
-    device = "cuda"
     torch.cuda.set_device(rank)
 
     epochs, batch, input_width, hidden, target_width = 5, 3, 20, 10, 5
@@ -474,14 +453,10 @@ def run_test_multiple_groups(rank, world_size, tempfile_name, backend, reduce_bu
 
 @skip_if_less_than_four_gpu
 @pytest.mark.parametrize("reduce_buffer_size", [0, 2 ** 20])
-@pytest.mark.parametrize("backend", ["gloo", "nccl"])
-def test_multiple_groups(reduce_buffer_size, backend):
+def test_multiple_groups(reduce_buffer_size):
     world_size = 4
     temp_file_name = tempfile.mkstemp()[1]
 
     mp.spawn(
-        run_test_multiple_groups,
-        args=(world_size, temp_file_name, backend, reduce_buffer_size),
-        nprocs=world_size,
-        join=True,
+        run_test_multiple_groups, args=(world_size, temp_file_name, reduce_buffer_size), nprocs=world_size, join=True,
     )